fix: remove workspace refresh when switching to code tab
There used to be a bug (scratchfoundation/scratch-gui#413) in the GUI
necessitating a manual refresh of the workspace. The underlying issue was dealt
with in scratch-blocks, as mentioned in scratchfoundation/scratch-gui#779.
However, the workspace refresh introduced in scratchfoundation/scratch-gui#459,
was never removed. This unnecessary call should be removed to improve
performance and to prevent unwanted PROJECT_CHANGED emissions.